+++ browser/CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog covers all five packages, as they are (for now) updated as a whol
 - [#850](https://github.com/atomicdata-dev/atomic-server/issues/850) Add drag & drop sorting to ResourceArray inputs.
 - [#757](https://github.com/atomicdata-dev/atomic-server/issues/757) Add drag & drop sorting to sidebar.
 - [#873](https://github.com/atomicdata-dev/atomic-server/issues/873) Add option to allow multiple resources in relation columns (Tables).
+- [#825](https://github.com/atomicdata-dev/atomic-server/issues/825) Folder display styles are now saved locally instead of on the resource. The display style property will now act as the default view style.
 
 ### @tomic/lib
 
@@ -44,6 +45,7 @@ This changelog covers all five packages, as they are (for now) updated as a whol
 - Added `useCollectionPage` hook.
 - Fix bug where `useCollection` would fetch the collection twice on mount.
 - `useServerURL` no longer stores the server url in localstorage.
+- BREAKING CHANGE: Removed the `useLocalStorage` hook.
 
 ### @tomic/cli